test(push): migrate dispatcher behavior coverage
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
